Refactor binary search algorithm to allow strategy to be passed in
This is the first step toward creating a binary search algorithm
that can search different ranges in different directions (i.e. when
going high on one range, go low on the other).

Since the :midpoint strategy requires some manipulation of the ranges
to be searched, stipulating these ranges had to be moved into the
strategy function as well.